dom/media/test/test_trackelementevent.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Bug 882677 - Implement the 'sourcing out of band text tracks' algorithm</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<video id="v" src="seek.webm" preload="auto">
<script type="text/javascript">

function runTest() {
  var video = document.getElementById("v");

  var trackOne = document.createElement("track");
  trackOne.src = "basic.vtt";
  trackOne.kind = "subtitles";

  var trackTwo = document.createElement("track");
  trackTwo.src = "bad-signature.vtt";
  trackTwo.kind = "captions";

  var trackThree = document.createElement("track");
  trackThree.src = "bad.vtt";
  trackThree.kind = "chapters";

  var events = 0;
  function trackOneEvent() {
    ok(true, "A load event for trackOne should have happened.");
    events++ && events == 3 && SimpleTest.finish();
  }
  function trackTwoEvent() {
    ok(true, "An error event for trackTwo should have happened.");
    events++ && events == 3 && SimpleTest.finish();
  }
  function trackThreeEvent() {
    ok(true, "An error event for trackThree should have happened.");
    events++ && events == 3 && SimpleTest.finish();
  }

  function shouldNotBeCalled() {
    ok(false, "Event should not have been called.");
  }

  trackOne.addEventListener("load", trackOneEvent);
  trackOne.addEventListener("error", shouldNotBeCalled)
  trackTwo.addEventListener("load", shouldNotBeCalled);
  trackTwo.addEventListener("error", trackTwoEvent);
  trackThree.addEventListener("load", shouldNotBeCalled);
  trackThree.addEventListener("error", trackThreeEvent);

  video.appendChild(trackOne);
  video.appendChild(trackTwo);
  video.appendChild(trackThree);
}

SimpleTest.waitForExplicitFinish();
onload = runTest;

</script>
</body>
</html>
